Finalized algorithm logic and added games to team path map
The code runs now and almost completes the bracket but errors out at the end because a there are a few team names that are in the path map that aren't in the five-thirty-eight data that I have loaded. Not sure why yet, hopefully we just missed an index or something.