Fixed team names, dealt with Wildcard teams and refined script
Many of the team names in team path map had abbrevs. that didn't line up with the 538 data. I changed them so that they all matched. We also changed the subset of the 538 data that we used so that we could exclude the 4 "wildcard" teams that didn't end up making the tournament (this was causing probability issues.

Problems persisting with np.random.choice() so I am experimenting with a custom version. Works in some cases, not final yet.